requirements.txt
Django==1.11.2
djangorestframework==3.6.3
numpy==1.13.0
olefile==0.44
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
svgwrite==1.1.11
